Minor code clean ups using the new code analysis stuff
Just trying out all the cool new goodness in MonoDevelop
git master (pre 3.1) and the newly added code analysis
options

Also, moving some of the test stuff to a test project
and adding a couple tests

+++ src/Clients/MainApp/FSpot.Widgets/FolderTreeModel.cs
@@ -200,7 +200,7 @@ namespace FSpot.Widgets
 					last_count += (int)GetValue (iter, 1);
 					SetValue (iter, 1, last_count);
 				}
-				count_all += (int)last_count;
+				count_all += last_count;
 			}
 		}
 	}
